scripts: builddeb: Allow specifying a package name
We want to be able to upgrade kernel packages.  In order to do this, we
need to have a consistent naming scheme, not one with the version number
embedded into the name.  Take a new parameter, KDEB_PKGNAME, and if specified
set the package name to that, rather than to the version number.

Signed-off-by: Sean Cross <xobs@kosagi.com>